# src/docs/swagger.yaml
basePath: /api/magic-circle
definitions:
...
  sector.CreateSectorReq:
    properties:
      coords:
        type: string
      description:
        type: string
    type: object
  sector.GetAllSectorsResp:
    properties:
      sectors:
        items:
          $ref: '#/definitions/sector.Sector'
        type: array
    type: object
  sector.Sector:
    properties:
      coords:
        type: string
      description:
        type: string
      id:
        type: integer
    type: object
  sector.UpdateSectorReq:
    properties:
      coords:
        type: string
      description:
        type: string
    type: object
info:
  contact: {}
  description: This is a server to get projects from github
  title: Magic-Circle API
  version: "1.0"
paths:
...
  /v1/sector:
    get:
      description: |-
        return all sectors
        quey params can make a logical predicates for example
        request: "/sectors?instutute=1+2&profile=1" equal "WHERE (institute_id=1 and profile_id=1) or institute_id=2"
      parameters:
      - description: institute name
        in: query
        name: institute
        type: string
      - description: direction name
        in: query
        name: direction
        type: string
      - description: profile name
        in: query
        name: profile
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sector.GetAllSectorsResp'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Sectors
      tags:
      - sector
    post:
      consumes:
      - application/json
      description: |-
        create sector according to giving coords
        coords is unique string
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/sector.CreateSectorReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sector.Sector'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create Sector
      tags:
      - sector
  /v1/sector/{id}:
    delete:
      description: delete sector and all adjacenttables that relate with this sector
      parameters:
      - description: id of sector
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete Sector
      tags:
      - sector
    put:
      consumes:
      - application/json
      description: update sector
      parameters:
      - description: id of sector
        in: path
        name: id
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/sector.UpdateSectorReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sector.Sector'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update Sector
      tags:
      - sector
  /v1/sectors:
    post:
      consumes:
      - application/json
      description: create sectors
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/sector.CreateSectorReq'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/sector.Sector'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create Sectors
      tags:
      - sector
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"